Changed dungeon and fight data structure
Dungeon and fight data are no longer in stored in Player but rather in
their dedicated classes. The classes do not define the parent Player, so
some methods (such as getReputation and setHp) will require additional
values that they used to fetch from the Player (such as XP and max HP).