Changed how allowed items are selected
You can now pass a predicate determining what items are allowed and what
aren't to inventory views. This is to make the item selection system
more flexible and to predetermine if an item can be used in the
InventoryChoiceDialog, rather that making the use method of the item do
it (it still does that though). The inventory view now also marks the
items that can be used by bolding them.